Moved PortDirection to its own file
