+++ lluvia/python/src/lluvia/core/enums/__init__.py
@@ -7,6 +7,6 @@
 """
 
 from .compute_dimension import ComputeDimension
-from .node import PortDirection, PortType
+from .node import PortType
 from .vulkan import Format, ImageType
 from .parameter import ParameterType
