+++ lluvia/python/src/lluvia/core/node_new/__init__.py
@@ -8,3 +8,4 @@
 
 from lluvia.core.node_new.node_state import *
 from lluvia.core.node_new.node_type import *
+from lluvia.core.node_new.port_direction import *
